Added account service REST endpoint
