+++ account-service/src/main/java/com/bandit/account/AccountServiceApplication.java
@@ -1,0 +1,24 @@
+package com.bandit.account;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.circuitbreaker.EnableCircuitBreaker;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.oauth2.config.annotation.web.configuration.EnableOAuth2Client;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+@EnableOAuth2Client
+@EnableFeignClients
+@EnableCircuitBreaker
+@EnableGlobalMethodSecurity(prePostEnabled = true)
+public class AccountServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(AccountServiceApplication.class, args);
+    }
+
+}
+
