+++ account-service/src/main/java/com/bandit/account/controller/AccountController.java
@@ -1,0 +1,11 @@
+package com.bandit.account.controller;
+
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+public class AccountController {
+    @RequestMapping(path = "/ping", method = RequestMethod.GET)
+    public String ping() {
+        return "OK";
+    }
+}
