--- account-service/src/main/java/com/bandit/accountservice/AccountServiceApplication.java
@@ -1,18 +1,0 @@
-package com.bandit.accountservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
-
-@SpringBootApplication
-@EnableDiscoveryClient
-@EnableZuulProxy
-public class AccountServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AccountServiceApplication.class, args);
-    }
-
-}
-
