+++ account-service/src/test/java/com/bandit/account/AccountServiceApplicationTests.java
@@ -1,15 +1,16 @@
-package com.bandit.accountservice;
+package com.bandit.account;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 
-
 @RunWith(SpringRunner.class)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@SpringBootTest
 public class AccountServiceApplicationTests {
+
     @Test
     public void contextLoads() {
     }
+
 }
