+++ account-service/src/test/resources/bootstrap.yml
@@ -1,4 +1,3 @@
-spring:
-  cloud:
-    config:
-      enabled: false
+eureka:
+  client:
+    enabled: false
